requirements.txt
beautifulsoup4 
selenium
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client

